Module D_state_option.StrictMode

Option for enabling/disabling strict mode engine.

include Accessor with type t = bool

type t = bool

The data saved in the state.

val get : D_loop.Typer.state -> t

Returns the option stored in the state. If it has not been registered, fetches the default option in the module Options.

val set : t -> D_loop.Typer.state -> D_loop.Typer.state

Sets the option on the dolmen state.

val clear : D_loop.Typer.state -> D_loop.Typer.state

Resets the option to its default value in Options.