Module type Domains_intf.EntryNotation

type entry

The type of entries that hold a mutable reference to a domain.

type domain

The type of domains.

val (!!) : entry -> domain

Return the domain associated with this entry.

val update : ex:Explanation.t -> entry -> domain -> unit

update ~ex e d updates the domain associated with e, intersecting it with d. The explanation ex is added to d.

raises Domain.Inconsistent
if the domains are incompatible.