Module Ephemeral.Canon

The Canon module first computes the canonical representative in an Uf.t instance before accessing the ephemeral map.

include EphemeralDomainMap with type key = X.r and type Entry.domain = domain

type t

The type of ephemeral domain maps, i.e. an imperative structure mapping keys to their current domain.

type key = X.r

The type of keys in the ephemeral map.

module Entry : sig ... end

val entry : t -> key -> Entry.t

entry t k returns the entry associated with k.

There is a unique entry associated with each key k that is created on-the-fly when entry t k is called for the first time. Calling entry t k with the same key will always return the same (physical) entry.

The domain associated with the entry is initialized from the underlying persistent domain (or the default function provided to edit) the first time it is accessed, and updated with set_domain or update.

include EntryNotation with type entry := Entry.t and type domain := domain

val (!!) : Entry.t -> domain

Return the domain associated with this entry.

val update : ex:Explanation.t -> Entry.t -> domain -> unit

update ~ex e d updates the domain associated with e, intersecting it with d. The explanation ex is added to d.

raises Domain.Inconsistent
if the domains are incompatible.