Module AltErgoLib.Intervals

val src : Logs.src

val map_bound : ('a -> 'b) -> 'a Intervals_intf.bound -> 'b Intervals_intf.bound

map_bound f b applies f to a finite (open or closed) bound b and does not change an unbounded bound.

This module provides implementations of union-of-intervals over reals and integers.

type 'a union

Polymorphic union type. This allows writing conversion functions between integer and real unions.

module Real : Intervals_intf.AlgebraicField with type explanation := Explanation.t and type value := Q.t and type 'a union = 'a union

Union-of-intervals over real numbers.

module Int : sig ... end

Union-of-intervals over integers.

module Legacy : sig ... end

The Legacy module reimplements (most of) the old legacy Intervals module on top of the current implementation to ease the transition.