Module AltErgoLib.My_unix

Unix wrapper

This module defines some wrappers around Unix function, in order to more easily maintain compatibility when compiling to javascript.

val cur_time : unit -> float

Returns the current time. *

val set_timeout : float -> unit

Set a timeout, using Unix timers. No-op on javascript.

val unset_timeout : unit -> unit

Unset the previously set timer.