Module AltErgoLib.My_zip

A wrapper of the Zip module of CamlZip: we use Zip except when we want to generate the.js file for try-Alt-Ergo *

type in_file

type entry

val open_in : string -> in_file

val close_in : in_file -> unit

val entries : in_file -> entry list

val read_entry : in_file -> entry -> string

val filename : entry -> string

val is_directory : entry -> bool

val extract_zip_file : string -> string

extract_zip_file filename extract the unique file of the zip archive filename and its content.