Module Alt_ergo_common.Parse_command

Parse_command module used at start-up to parse the command line

val parse_cmdline_arguments : unit -> unit

Only exported function. Calling it will parse the command line options and set them accordingly for the rest of the execution

exception Exit_parse_command of int

Exception used to exit with corresponding retcode